+++ fractal_server/app/routes/admin/v2/task_group_reset.py
@@ -5,6 +5,7 @@ from fastapi import HTTPException
 from fastapi import Response
 from fastapi import status
 from pydantic import BaseModel
+from pydantic import ConfigDict
 from pydantic import Field
 from pydantic import Json
 
@@ -25,6 +26,7 @@ from fractal_server.app.routes.auth import current_superuser_act
 from fractal_server.app.routes.aux._python_interpreter import (
     get_python_interpreter_or_422,
 )
+from fractal_server.app.routes.aux.pixi_version import get_pixi_version_or_422
 from fractal_server.app.routes.aux.validate_user_profile import (
     validate_user_profile,
 )
@@ -105,6 +107,7 @@ class TaskGroupOverridesPip(BaseModel):
         pinned_package_versions_post:
     """
 
+    model_config = ConfigDict(extra="forbid")
     pip_extras: str | None = None
     python_version: NonEmptyStr | None = None
     pinned_package_versions_pre: Json[dict[NonEmptyStr, NonEmptyStr]] = Field(
@@ -113,6 +116,19 @@ class TaskGroupOverridesPip(BaseModel):
     pinned_package_versions_post: Json[dict[NonEmptyStr, NonEmptyStr]] = Field(
         default_factory=dict
     )
+
+
+class TaskGroupOverridesPixi(BaseModel):
+    """
+    Overrides of the original task-group properties.
+
+    Attributes:
+        use_pixi_lockfile:
+    """
+
+    model_config = ConfigDict(extra="forbid")
+
+    use_pixi_lockfile: bool = False
 
 
 async def _create_activity(
@@ -210,17 +226,6 @@ async def reset_tasks_pip(
     return task_group_activity
 
 
-class TaskGroupOverridesPixi(BaseModel):
-    """
-    Overrides of the original task-group properties.
-
-    Attributes:
-        use_pixi_lockfile:
-    """
-
-    use_pixi_lockfile: bool = False
-
-
 @router.post(
     "/{task_group_id}/reset/pixi/",
     response_model=TaskGroupActivityRead,
@@ -245,6 +250,11 @@ async def reset_tasks_pixi(
     resource, profile = await validate_user_profile(user=owner, db=db)
     _verify_support(
         pip_or_pixi="pixi", task_group=task_group, resource=resource
+    )
+
+    get_pixi_version_or_422(
+        pixi_version=task_group.pixi_version,
+        resource=resource,
     )
 
     logger.info(